done4 .yes
n this example, you will learn to check whether a number entered by the user is even or odd.

To understand this example, you should have the knowledge of the following Python programming topics:

Python Operators
Python if...else Statement
A number is even if it is perfectly divisible by 2. When the number is divided by 2, we use the remainder operator % to compute the remainder. If the remainder is not zero, the number is odd.

Source Code